requirements.txt
boto3==1.37.27
torch==2.6.0
pillow==11.1.0
transformers==4.50.3
fastapi==0.115.12
uvicorn==0.34.0
